backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
reportlab==4.0.7
openai==1.3.5
groq==0.37.1
python-dotenv==1.0.0


